chore: renames internal package from docker to provider
