+++ internal/provider/config.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"crypto/tls"
