+++ internal/provider/data_source_docker_registry_image.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"crypto/sha256"
