+++ internal/provider/label.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"fmt"
