+++ internal/provider/label_migration_test.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"fmt"
