+++ internal/provider/provider_test.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"os/exec"
