+++ internal/provider/resource_docker_config.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"context"
