+++ internal/provider/resource_docker_container.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"log"
