+++ internal/provider/resource_docker_container_migrate.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"fmt"
