+++ internal/provider/resource_docker_container_v1.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
