+++ internal/provider/resource_docker_image.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
