+++ internal/provider/resource_docker_image_funcs.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"bytes"
