+++ internal/provider/resource_docker_network.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"log"
