+++ internal/provider/resource_docker_registry_image.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"os"
