+++ internal/provider/resource_docker_registry_image_funcs.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"archive/tar"
