+++ internal/provider/resource_docker_service_funcs.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"context"
