+++ internal/provider/resource_docker_volume.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"context"
