+++ internal/provider/structures_service.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"strconv"
