+++ internal/provider/validators.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import (
 	"encoding/base64"
