+++ internal/provider/validators_test.go
@@ -1,4 +1,4 @@
-package docker
+package provider
 
 import "testing"
